get_energy_loss move to a new module to solve import loops
